+++ CHANGELOG.md
@@ -6,7 +6,10 @@
 
 - Support `mg new` to create `Miragefile.secret` at once.
 
+- Support `mg new` to create `LICENSE` doc at once.
+
 - Introduce `mg ? system` to check OS and Python version.
+
 
 # `latest release` v0.1.4